Do not run new estimator test for native estimator
